Int overflow in the datagram timeout

+++ javasdk/NRSDK/src/com/neuronrobotics/sdk/common/BowlerDatagramFactory.java
@@ -29,7 +29,7 @@ public class BowlerDatagramFactory {
 	private static BowlerDatagram pool [];
 	private static int failed=0;
 	private static int poolDefaultSize = 1500;
-	private static int packetTimeout = 2000;
+	private static long packetTimeout = 2000;
 	
 	static{
 		if(instance ==  null){
@@ -219,7 +219,7 @@ public class BowlerDatagramFactory {
 		BowlerDatagramFactory.poolDefaultSize = poolSize;
 	}
 
-	public static int getPacketTimeout() {
+	public static long getPacketTimeout() {
 		return packetTimeout;
 	}
